Changes Made:
- Made some minor changes on the cascade type of relationship of entities

+++ src/main/java/com/ims/inventory_management_system_backend/entities/purchase_items/PurchaseItems.java
@@ -24,7 +24,7 @@ public class PurchaseItems {
     private Integer quantity;
 
     @Column(name = "unit_cost", nullable = false)
-    private Integer unitCost;
+    private Double unitCost;
 
     @Column(name = "discount", nullable = false)
     private Double discount;
